Refactor: 로그인 공통 로직 외의 기능 수정 및 테스트
